parsing-service/requirements.txt
# requirements.txt
flask
google-cloud-storage
google-cloud-aiplatform
PyMuPDF
gunicorn